Fix homebrew checks when it's not installed

+++ shells/zsh/settings.zsh
@@ -59,9 +59,11 @@ zstyle ':completion:*' format '%B---- %d%b'
 # This is also necessary for hub's Zsh completions to work:
 # https://github.com/github/hub/issues/1956.
 function () {
-	GIT_ZSH_COMPLETIONS_FILE_PATH="$(brew --prefix)/share/zsh/site-functions/_git"
-	if [ -f $GIT_ZSH_COMPLETIONS_FILE_PATH ]; then
-		rm $GIT_ZSH_COMPLETIONS_FILE_PATH
+	if command -v brew >/dev/null 2>&1; then
+		GIT_ZSH_COMPLETIONS_FILE_PATH="$(brew --prefix)/share/zsh/site-functions/_git"
+		if [ -f $GIT_ZSH_COMPLETIONS_FILE_PATH ]; then
+			rm $GIT_ZSH_COMPLETIONS_FILE_PATH
+		fi
 	fi
 }
